Move models to `abstract_models` to facilitate model customization
