+++ plans_i18n/translation.py
@@ -1,5 +1,10 @@
+from django.db.models import get_model
 from modeltranslation.translator import translator, TranslationOptions
 from plans.models import Plan, Pricing, Quota
+
+Plan = get_model('plans', 'Plan')
+Pricing = get_model('plans', 'Pricing')
+Quota = get_model('plans', 'Quota')
 
 
 # Translations for django-plans
